--- src/components/TodoContext.jsx
@@ -1,5 +1,0 @@
-import React, { createContext } from "react";
-
-const TodoContext = createContext();
-
-export default TodoContext;
