+++ src/components/TodoList.jsx
@@ -2,7 +2,7 @@ import React, { useContext } from "react";
 import styled from "styled-components";
 import Li from "../atomicComponents/Li";
 import XButton from "../atomicComponents/XButton";
-import TodoContext from "./TodoContext";
+import { TodoContext } from "./TodoContextProvider";
 
 const StyledLi = styled(Li)`
   text-decoration: ${({ textDeco }) => (textDeco === "done" ? "line-through" : "none")};
